Add any already published photos to the list returned from the collections API when syncing collections.
The Collections API sometimes doesn't return all of the photos in a collection (https://github.com/500px/api-documentation/issues/83).
This can mean that recently uploaded photos are dropped from collections if another upload or delete is performed on the same collection.
Adding the list of already published photos to the list returned from the API solves this though will re-add photos to the collection
if the user had deleted them through the website.