Remove the plugin code
The 500px Lightroom plugin is no longer supported by 500px. This means
we’re not currently making updates to it, and our Support Team is no
longer available for assistance or troubleshooting with any version you
may have access to.

This also means that we have removed the ability to download it from our
site. We caution the usage of the 500px Lightroom Plugin or downloading
it from any other sites since it is now obsolete.